requirements-dev.txt
pytest>=8.0.0
paho-mqtt>=2.1.0
requests>=2.32.0
